renamed DeleteDialogComponent
